requirements.txt
asgiref==3.5.2
Django==4.1
django-environ==0.7.0
djangorestframework==3.13.1
greenlet==1.1.2
psycopg2==2.9.3
pytz==2022.2.1
SQLAlchemy==1.4.40
sqlparse==0.4.2
tzdata==2022.2
